revert: Moves ClientConnectionServer and WebsocketServer back to presentation layer.
